ps pro working gemini - whisper - Copy/requirements.txt
google-generativeai
markdownify==0.11.6
openai>=1.0.0
Pillow==10.2.0
pyaudio
python-dotenv==1.0.0
python-wordpress-xmlrpc==2.3
requests==2.31.0
speechrecognition
streamlit==1.32.0
tabulate
urllib3
whisper
--extra-index-url https://download.pytorch.org/whl/cu118
